Backend/requirements.txt
fastapi
uvicorn
openai
psycopg2-binary
pyjwt
requests
beautifulsoup4
